+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Upgrade CCS Frontend Helpers version to v2.4.0 ([PR 4821](https://github.com/Crown-Commercial-Service/crown-marketplace/pull/4821))
 - Upgrade CCS Frontend version to v1.4.1 ([PR 4821](https://github.com/Crown-Commercial-Service/crown-marketplace/pull/4821))
 - Upgrade ruby version to v3.4.3 ([PR 4835](https://github.com/Crown-Commercial-Service/crown-marketplace/pull/4835))
+- Upgrade Rails to v8.0.2.0 ([PR 4858](https://github.com/Crown-Commercial-Service/crown-marketplace/pull/4858))
 
 ## [6.2.0] - 2025-01-23